restructure ok
