+++ config/kaanh-config.cmake.in
@@ -1,0 +1,14 @@
+@PACKAGE_INIT@ 
+
+set(RELY_LINK_DIR "@RELY_LINK_DIR@")
+link_directories(${RELY_LINK_DIR})
+
+set_and_check(kaanh_INCLUDE_DIRS "@kaanh_INCLUDE_DIRS@")
+set(kaanh_LIBRARIES "@kaanh_LIBRARIES@")
+if(UNIX)
+	set(kaanh_LIBRARIES -Wl,--start-group ${aris_LIBRARIES} -Wl,--end-group)
+endif(UNIX)
+
+
+include("${CMAKE_CURRENT_LIST_DIR}/kaanh-lib-release.cmake")
+include("${CMAKE_CURRENT_LIST_DIR}/kaanh-lib-debug.cmake")
