+++ huaerkang/CMakeLists.txt
@@ -1,6 +1,6 @@
 ﻿cmake_minimum_required(VERSION 3.12)
 
-project(kaanh)
+project(huaerkang)
 
 # aris use c++ 17 standard
 set(CMAKE_CXX_STANDARD 17)
@@ -15,28 +15,21 @@ IF(WIN32)
 endif(WIN32)
 include_directories(${aris_INCLUDE_DIRS})
 
+# find kaanh
 if(UNIX)
-	include_directories(usr/include/kaanh)
-endif(UNIX)
-
-if(WIN32)
-	link_directories(D:/1working/build_aris_kaanh/Debug)
-endif(WIN32)
-
-message(${aris_DIR})
-add_executable(kaanh src/main.cpp src/kaanh.h src/config.cpp src/config.h src/kaanhconfig.h)
-
-if(UNIX)
-	target_link_libraries(kaanh libkaanh_lib.a)
-endif(UNIX)
-
-IF(WIN32)
-	target_link_libraries(kaanh kaanh_lib)
-endif(WIN32)
-
-if(UNIX)
-    target_link_libraries(kaanh ${aris_LIBRARIES} atomic)
+	add_definitions(-DUNIX)
+        find_package(kaanh REQUIRED PATHS /usr/kaanh)
 endif(UNIX)
 IF(WIN32)
-	target_link_libraries(kaanh ${aris_LIBRARIES})
+        find_package(kaanh REQUIRED PATHS C:/kaanh/kaanh-1.0)
+endif(WIN32)
+include_directories(${kaanh_INCLUDE_DIRS})
+
+message(${aris_DIR} ${kaanh_DIR})
+add_executable(huaerkang src/main.cpp src/kaanh.h src/config.cpp src/config.h src/kaanhconfig.h)
+if(UNIX)
+    target_link_libraries(huaerkang ${aris_LIBRARIES} ${kaanh_LIBRARIES} atomic)
+endif(UNIX)
+IF(WIN32)
+	target_link_libraries(huaerkang ${aris_LIBRARIES} ${kaanh_LIBRARIES})
 endif(WIN32)
